Excercise.py
This is a program for creating the dictionary. In this you will going to give a word as the input and if that word is present in the dictionary which is created by me then the meaning of that word will get printed else, an exception will be shown to you, which is also handled my me in the program.